Rename CaseService to CaseClient.
